refactor(agent): revert some abstractions
